エラー修正

+++ hosihime/hosihime/Source/TimeScore.cpp
@@ -12,7 +12,7 @@ void TimeScore::init()
 }
 void TimeScore::setScore(Timer& timer)
 {
-	score=(timer.getDefTime()*FRAMETIME)-timer.getTime();
+	score=((int)(timer.getDefTime()*FRAMETIME))-timer.getTime();
 }
 const int TimeScore::getScore() const 
 {
